requirements.txt
flask
gunicorn
google.generativeai
textblob
markdown
re
numpy